// src/openapi/models/FeedTypes.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Baby Tracker
 * Track everything your baby does
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * An enumeration.
 * @export
 * @enum {string}
 */
export enum FeedTypes {
    NUMBER_1 = 1,
    NUMBER_2 = 2,
    NUMBER_3 = 3,
    NUMBER_4 = 4,
    NUMBER_5 = 5
}

export function FeedTypesFromJSON(json: any): FeedTypes {
    return FeedTypesFromJSONTyped(json, false);
}

export function FeedTypesFromJSONTyped(json: any, ignoreDiscriminator: boolean): FeedTypes {
    return json as FeedTypes;
}

export function FeedTypesToJSON(value?: FeedTypes | null): any {
    return value as any;
}